added package matching to benchmark
