+++ src/commands/search.rs
@@ -10,30 +10,11 @@ use regex::Regex;
 use sfsu::{
     buckets::Bucket,
     output::sectioned::{Children, Section, Sections, Text},
-    packages::manifest::StringOrArrayOfStringsOrAnArrayOfArrayOfStrings,
+    packages::{manifest::StringOrArrayOfStringsOrAnArrayOfArrayOfStrings, SearchMode},
 };
 
 use sfsu::packages::{is_installed, Manifest};
 use strum::Display;
-
-#[derive(Debug, Default, Copy, Clone, ValueEnum, Display, Parser)]
-#[strum(serialize_all = "snake_case")]
-enum SearchMode {
-    #[default]
-    Name,
-    Binary,
-    Both,
-}
-
-impl SearchMode {
-    pub fn match_names(self) -> bool {
-        matches!(self, SearchMode::Name | SearchMode::Both)
-    }
-
-    pub fn match_binaries(self) -> bool {
-        matches!(self, SearchMode::Binary | SearchMode::Both)
-    }
-}
 
 #[derive(Debug, Clone, Parser)]
 pub struct Args {
@@ -55,135 +36,6 @@ pub struct Args {
 
     #[clap(short, long, help = "Search mode to use", default_value_t)]
     mode: SearchMode,
-}
-
-#[derive(Debug, Clone)]
-struct MatchCriteria {
-    name: bool,
-    bins: Option<Vec<String>>,
-}
-
-impl MatchCriteria {
-    pub const fn new() -> Self {
-        Self {
-            name: false,
-            bins: None,
-        }
-    }
-
-    pub fn matches(
-        file_name: &str,
-        manifest: Option<&Manifest>,
-        mode: SearchMode,
-        pattern: &Regex,
-    ) -> Self {
-        let file_name = file_name.to_string();
-
-        let mut output = MatchCriteria::new();
-
-        if mode.match_names() && pattern.is_match(&file_name) {
-            output.name = true;
-        }
-
-        if let Some(manifest) = manifest {
-            let binaries = manifest
-                .bin
-                .clone()
-                .map(StringOrArrayOfStringsOrAnArrayOfArrayOfStrings::to_vec)
-                .unwrap_or_default();
-
-            let binary_matches = binaries
-                .into_iter()
-                .filter(|binary| pattern.is_match(binary))
-                .filter_map(|b| {
-                    if pattern.is_match(&b) {
-                        Some(b.clone())
-                    } else {
-                        None
-                    }
-                })
-                .collect_vec();
-
-            output.bins = Some(binary_matches);
-        }
-
-        output
-    }
-}
-
-fn parse_output(
-    manifest: &Manifest,
-    bucket: impl AsRef<str>,
-    installed_only: bool,
-    pattern: &Regex,
-    mode: SearchMode,
-) -> Option<Section<Text<String>>> {
-    // TODO: Better display of output
-
-    // This may be a bit of a hack, but it works
-
-    let match_output = MatchCriteria::matches(
-        &manifest.name,
-        if mode.match_binaries() {
-            Some(manifest)
-        } else {
-            None
-        },
-        mode,
-        pattern,
-    );
-
-    if !match_output.name && match_output.bins.is_none() {
-        return None;
-    }
-
-    // TODO: Refactor to remove pointless binary matching on name-only search mode
-    // TODO: Fix error parsing manifests
-
-    let is_installed = is_installed(&manifest.name, Some(bucket));
-    if installed_only && !is_installed {
-        return None;
-    }
-
-    let styled_package_name = if manifest.name == pattern.to_string() {
-        manifest.name.bold().to_string()
-    } else {
-        manifest.name.clone()
-    };
-
-    let installed_text = if is_installed && !installed_only {
-        "[installed] "
-    } else {
-        ""
-    };
-
-    let title = format!(
-        "{styled_package_name} ({}) {installed_text}",
-        manifest.version
-    );
-
-    let package = if mode.match_binaries() {
-        let bins = if let Some(bins) = match_output.bins {
-            bins.iter()
-                .map(|output| {
-                    Text::new(format!(
-                        "{}{}\n",
-                        sfsu::output::sectioned::WHITESPACE,
-                        output.bold()
-                    ))
-                })
-                .collect_vec()
-        } else {
-            vec![]
-        };
-
-        Section::new(Children::Multiple(bins))
-    } else {
-        Section::new(Children::None)
-    }
-    .with_title(title);
-
-    Some(package)
 }
 
 impl super::Command for Args {
@@ -227,7 +79,7 @@ impl super::Command for Args {
                 let matches = bucket_contents
                     .par_iter()
                     .filter_map(|file| {
-                        parse_output(file, bucket.name(), self.installed, &pattern, self.mode)
+                        file.parse_output(bucket.name(), self.installed, &pattern, self.mode)
                     })
                     .collect::<Vec<_>>();
